drop_data_tables function improved - if db file not exists

+++ dbhandling/prozorro_db_create.py
@@ -211,40 +211,41 @@ def drop_data_tables(db_filename):
     load_dotenv()
     db_fn = os.path.join(".", os.getenv("PROZORRO_DB_FOLDER"), db_filename)
 
-    try:
-        connection_obj = sqlite3.connect(db_fn)
-        connection_obj.execute("PRAGMA foreign_keys=0;")
-        cursor_obj = connection_obj.cursor()
+    if os.path.isfile(db_fn):
+        try:
+            connection_obj = sqlite3.connect(db_fn)
+            connection_obj.execute("PRAGMA foreign_keys=0;")
+            cursor_obj = connection_obj.cursor()
 
-        table = "DROP TABLE IF EXISTS items;"
-        cursor_obj.execute(table)
-        connection_obj.commit()
+            table = "DROP TABLE IF EXISTS items;"
+            cursor_obj.execute(table)
+            connection_obj.commit()
 
-        table = "DROP TABLE IF EXISTS procuringEntity;"
-        cursor_obj.execute(table)
-        connection_obj.commit()
+            table = "DROP TABLE IF EXISTS procuringEntity;"
+            cursor_obj.execute(table)
+            connection_obj.commit()
 
-        table = "DROP TABLE IF EXISTS Tender;"
-        cursor_obj.execute(table)
-        connection_obj.commit()
+            table = "DROP TABLE IF EXISTS Tender;"
+            cursor_obj.execute(table)
+            connection_obj.commit()
 
-        table = "DROP TABLE IF EXISTS Tender_list;"
-        cursor_obj.execute(table)
-        connection_obj.commit()
+            table = "DROP TABLE IF EXISTS Tender_list;"
+            cursor_obj.execute(table)
+            connection_obj.commit()
 
-        connection_obj.execute("PRAGMA foreign_keys=1;")
-        connection_obj.commit()
+            connection_obj.execute("PRAGMA foreign_keys=1;")
+            connection_obj.commit()
 
-    except sqlite3.Error as e:
-        print(f"SQLite error {e.args[0]}")
-        telegram_channel_scripting.raise_tech_message(telegram_bot_token=os.getenv("TELEGRAM_BOT_TOKEN"),
-                                                      err_module="Prozorro Local DB",
-                                                      err_message="Database dropping error in " + db_fn + "." +
-                                                                  str(e.args[0]))
-        error_code = -80
+        except sqlite3.Error as e:
+            print(f"SQLite error {e.args[0]}")
+            telegram_channel_scripting.raise_tech_message(telegram_bot_token=os.getenv("TELEGRAM_BOT_TOKEN"),
+                                                          err_module="Prozorro Local DB",
+                                                          err_message="Database dropping error in " + db_fn + "." +
+                                                                      str(e.args[0]))
+            error_code = -80
 
-    finally:
-        if connection_obj:
-            connection_obj.close()
+        finally:
+            if connection_obj:
+                connection_obj.close()
 
     return error_code
